+++ src/main/java/org/example/delivery/store/model/dto/response/GetStoreResponse.java
@@ -1,0 +1,5 @@
+package org.example.delivery.store.model.dto.response;
+
+public record GetStoreResponse (
+
+){}
