--- src/main/java/org/example/delivery/store/model/response/StoreResponse.java
@@ -1,5 +1,0 @@
-package org.example.delivery.store.model.response;
-
-public record StoreResponse (
-    String message
-){}
