+++ src/main/java/org/example/delivery/store/service/StoreService.java
@@ -7,8 +7,8 @@ import org.example.delivery.common.domain.Store;
 import org.example.delivery.common.domain.User;
 import org.example.delivery.common.exception.ErrorCode;
 import org.example.delivery.common.exception.base.NotFoundException;
-import org.example.delivery.store.model.request.StoreRequest;
-import org.example.delivery.store.model.response.StoreResponse;
+import org.example.delivery.store.model.dto.request.StoreRequest;
+import org.example.delivery.store.model.dto.response.StoreResponse;
 import org.example.delivery.store.repository.StoreRepository;
 import org.example.delivery.user.repository.UserRepository;
 import org.springframework.stereotype.Service;
